added an option to automatically set embed size via fastai's rule of thumb. Re-structure the preprocessing module, breaking it up in a smaller modules for better debugging and 'tractability'
