+++ pytorch_widedeep/preprocessing/__init__.py
@@ -1,1 +1,4 @@
-from pytorch_widedeep.preprocessing.preprocessors import *
+from pytorch_widedeep.preprocessing.tab_preprocessor import TabPreprocessor
+from pytorch_widedeep.preprocessing.text_preprocessor import TextPreprocessor
+from pytorch_widedeep.preprocessing.wide_preprocessor import WidePreprocessor
+from pytorch_widedeep.preprocessing.image_preprocessor import ImagePreprocessor
